Use JTS GeometryFixer
